Move decoding helpers into decode.ts
