+++ packages/iov-bns/src/types.ts
@@ -1,4 +1,3 @@
-import * as Long from "long";
 import { As } from "type-tagger";
 
 import {
@@ -16,8 +15,6 @@ import {
   SwapClaimTransaction,
   SwapOfferTransaction,
 } from "@iov/bcp";
-
-import { Int53 } from "@iov/encoding";
 
 // config (those are not used outside of @iov/bns)
 
@@ -96,40 +93,6 @@ export interface Keyed {
 
 export interface Decoder<T extends {}> {
   readonly decode: (data: Uint8Array) => T;
-}
-
-/**
- * Decodes a protobuf int field (int32/uint32/int64/uint64) into a JavaScript
- * number.
- */
-export function asIntegerNumber(maybeLong: Long | number | null | undefined): number {
-  if (!maybeLong) {
-    return 0;
-  } else if (typeof maybeLong === "number") {
-    if (!Number.isInteger(maybeLong)) {
-      throw new Error("Number is not an integer.");
-    }
-    return maybeLong;
-  } else {
-    return maybeLong.toInt();
-  }
-}
-
-export function asInt53(input: Long | number | null | undefined): Int53 {
-  if (!input) {
-    return new Int53(0);
-  } else if (typeof input === "number") {
-    return new Int53(input);
-  } else {
-    return Int53.fromString(input.toString());
-  }
-}
-
-export function ensure<T>(maybe: T | null | undefined, msg?: string): T {
-  if (maybe === null || maybe === undefined) {
-    throw new Error("missing " + (msg || "field"));
-  }
-  return maybe;
 }
 
 // transactions
