+++ packages/iov-bns/types/decode.d.ts
@@ -1,6 +1,15 @@
+import * as Long from "long";
 import { Amount, ChainId, FullSignature, Nonce, PubkeyBundle, SignatureBytes, SignedTransaction, Token, UnsignedTransaction } from "@iov/bcp";
+import { Int53 } from "@iov/encoding";
 import * as codecImpl from "./generated/codecimpl";
 import { BnsUsernameNft, CashConfiguration, Keyed, Participant, PrivkeyBundle } from "./types";
+/**
+ * Decodes a protobuf int field (int32/uint32/int64/uint64) into a JavaScript
+ * number.
+ */
+export declare function asIntegerNumber(maybeLong: Long | number | null | undefined): number;
+export declare function asInt53(input: Long | number | null | undefined): Int53;
+export declare function ensure<T>(maybe: T | null | undefined, msg?: string): T;
 export declare function decodeUsernameNft(nft: codecImpl.username.IUsernameToken, registryChainId: ChainId): BnsUsernameNft;
 export declare function decodeNonce(acct: codecImpl.sigs.IUserData & Keyed): Nonce;
 export declare function decodePubkey(publicKey: codecImpl.crypto.IPublicKey): PubkeyBundle;
